+++ trackAndTrace/sponsored-transaction-service/src/types.rs
@@ -9,6 +9,7 @@ use concordium_rust_sdk::{
         Timestamp,
     },
     types::{smart_contracts::ExceedsParameterSize, Nonce, RejectReason, WalletAccount},
+    v2::upward::UnknownDataError,
 };
 use hex::FromHexError;
 use std::{
@@ -77,8 +78,8 @@ pub enum ServerError {
     /// The contract is not allowed to be used by the service.
     #[error("Contract address is not allowed to be used by the service: {contract}.")]
     ContractNotAllowed { contract: ContractAddress },
-    #[error("The type `${0}` is unkown to this SDK. This can happen if the SDK is not fully compatible with the Concordium node. You might want to update the SDK to a newer version.")]
-    Unknown(String),
+    #[error("UnknownDataError occured: {0}")]
+    UnknownDataError(#[from] UnknownDataError),
 }
 
 impl axum::response::IntoResponse for ServerError {